webapp/requirements/dev.txt
-r base.txt
django-debug-toolbar==1.5
bpython==0.15
